more updates for 0.5.0
