+++ .github/workflows/create-js-release.yaml
@@ -175,7 +175,7 @@ jobs:
         uses: actions/download-artifact@v2
         with:
           name: bindings-x86_64-apple-darwin
-          path: nodejs-polars
+          path: nodejs-polars/polars
       - name: Run test
         run: yarn test
   publish:
